refactor: separated the functions in sub-categories for readability
