+++ src/types.ts
@@ -1,0 +1,6 @@
+export type Todo = {
+  title: string
+  date: string
+  dateColor: string
+  checkbox: boolean
+}
